# openapi/v1.yml
openapi: '3.1.0'
info:
  title: TwextHub API
  version: '1.0.0'
  license:
    name: Apache-2.0
    identifier: Apache-2.0
  description: >
    TwextHub is a lightweight registry for publishing and discovering
    Twext-compiled TurboWarp extensions ("Twexts"). Packages are namespaced
    to their publishing account (`@namespace/id`) and versioned with SemVer.
    Publishing sends a gzipped tarball of the Twext project (`twext.yml` plus
    sources); the hub validates the manifest and compiles the extension
    itself in a sandboxed build, so the reviewed source is exactly what
    produces the served blob.

    The `/v1` surface documented here is stable as of 1.0.0. Breaking changes
    require a new major version; additive changes bump the minor (see
    CHANGELOG.md).

    Every authenticated (`Userspace`/`Adminspace`) endpoint requires the
    caller to have accepted the current Terms of Service; a caller who
    hasn't gets a `403` (see `Forbidden`) until they patch their own user with
    `termsAcceptedVersion` set to the `version` that `GET /terms` reports. A
    stale version is a `422`, so an acceptance is always recorded against text
    the account was shown. Exemptions that stay available without
    re-acceptance: `DELETE /sessions/current`, `GET /me`, the acceptance
    itself, password-only user updates, and account deletion.

    Bearer tokens come in two kinds. **Sessions** are created by
...
    every scope. **Automation tokens** are created explicitly via
...

...

...
servers:
...
security:
  - BearerAuth: []
paths:
...
  /orgs:
    post:
      summary: Create an organization
      operationId: createOrganization
      tags: [Userspace]
      description: >
        An organization is a pseudo-account: it owns a namespace, a profile and
        an extension collection, but has no password and cannot sign in. The
        caller becomes its first owner, which is the only reason it exists — an
        organization with no owner could not be changed by anyone.

        The namespace is shared with accounts, so a name already used by an
        account is a `409`. Creation is rate limited like a signup and needs the
        current Terms of Service accepted, like a publish.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [namespace]
              properties:
                namespace: { $ref: '#/components/schemas/Namespace' }
                displayName: { type: string, maxLength: 80 }
                bio: { type: ['string', 'null'], maxLength: 280 }
                website: { type: ['string', 'null'], maxLength: 400, format: uri }
                github:
                  type: ['string', 'null']
                  description: GitHub username, shown as a profile link.
                avatarUrl: { type: ['string', 'null'], maxLength: 400, format: uri }
                bannerUrl: { type: ['string', 'null'], maxLength: 400, format: uri }
      responses:
        '201':
          description: Organization created
          headers:
            Location:
              description: The new organization's own URL
              schema: { type: string }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Organization' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/UnprocessableEntity' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
    get:
      summary: List organizations
      operationId: listOrganizations
      tags: [Open]
      security: []
      parameters:
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Dir'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Paginated organizations
          content:
            application/json:
              schema:
                type: object
                required: [data, _links]
                properties:
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/Organization' }
                  _links: { $ref: '#/components/schemas/PageLinks' }
  /orgs/{namespace}:
    parameters:
      - $ref: '#/components/parameters/NamespacePath'
    get:
      summary: Get an organization
      operationId: getOrganization
      tags: [Open]
      security: []
      responses:
        '200':
          description: Organization found
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Organization' }
        '404': { $ref: '#/components/responses/NotFound' }
    patch:
      summary: Update an organization
      operationId: updateOrganization
      tags: [Userspace]
      description: >
        Any owner, or an admin. Only the profile fields exist here: an
        organization has no password, no role and no terms of its own, and its
        members are changed through `/orgs/{namespace}/owners`.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              minProperties: 1
              properties:
                displayName: { type: string, maxLength: 80 }
                bio: { type: ['string', 'null'], maxLength: 280 }
                website: { type: ['string', 'null'], maxLength: 400, format: uri }
                github:
                  type: ['string', 'null']
                  description: GitHub username, shown as a profile link.
                avatarUrl: { type: ['string', 'null'], maxLength: 400, format: uri }
                bannerUrl: { type: ['string', 'null'], maxLength: 400, format: uri }
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Organization' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '422': { $ref: '#/components/responses/UnprocessableEntity' }
    delete:
      summary: Delete an organization
      operationId: deleteOrganization
      tags: [Userspace]
      description: >
        Any owner, or an admin. Cascades to every extension, version, image and
        webhook under the namespace, and to the owner rows.
      responses:
        '204': { description: Deleted }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /orgs/{namespace}/owners:
    parameters:
      - $ref: '#/components/parameters/NamespacePath'
    get:
      summary: List an organization's owners
      operationId: listOrganizationOwners
      tags: [Open]
      security: []
      responses:
        '200':
          description: The accounts that manage this organization
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/OrganizationOwner' }
  /orgs/{namespace}/owners/{ownerNamespace}:
    parameters:
      - $ref: '#/components/parameters/NamespacePath'
      - name: ownerNamespace
        in: path
        required: true
        schema: { $ref: '#/components/schemas/Namespace' }
        description: The account to add or remove. An organization cannot own another organization.
    put:
      summary: Add an owner
      operationId: addOrganizationOwner
      tags: [Userspace]
      description: >
        Any owner, or an admin. Idempotent: adding an account that is already an
        owner is a `204` rather than an error, so a retry does not need to know
        whether the first attempt landed. The account is notified.
      responses:
        '204': { description: Added, or already an owner }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '422': { $ref: '#/components/responses/UnprocessableEntity' }
    delete:
      summary: Remove an owner
      operationId: removeOrganizationOwner
      tags: [Userspace]
      description: >
        Any owner, or an admin. The last owner cannot be removed — an
        organization nobody owns has no way back, so the row is either handed on
        or the organization deleted. The account is notified.
      responses:
        '204': { description: Removed }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
  /orgs/{namespace}/extensions:
    parameters:
      - $ref: '#/components/parameters/NamespacePath'
    get:
      summary: List an organization's extensions
      operationId: listOrganizationExtensions
      tags: [Open]
      security: []
      description: >
        The registry listing scoped to this namespace, with the same `sort`,
        `license`, cursor and paging behaviour. An owner of the organization
        sees its private extensions; everyone else sees only the public ones.
      parameters:
        - $ref: '#/components/parameters/Sort'
        - $ref: '#/components/parameters/LicenseFilter'
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Dir'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Paginated extensions
          content:
            application/json:
              schema:
                type: object
                required: [data, _links]
                properties:
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/ExtensionSummary' }
                  _links: { $ref: '#/components/schemas/PageLinks' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '404': { $ref: '#/components/responses/NotFound' }
  /orgs/{namespace}/webhooks:
    parameters:
      - $ref: '#/components/parameters/NamespacePath'
    get:
      summary: List an organization's webhooks
      operationId: listOrganizationWebhooks
      tags: [Userspace]
      description: >
        Any owner, or an admin. Secrets are never returned here; they are
        returned once, by the create call.
      responses:
        '200':
          description: The namespace-wide webhooks
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/Webhook' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
    post:
      summary: Create an organization webhook
      operationId: createOrganizationWebhook
      tags: [Userspace]
      description: >
        Any owner, or an admin. A webhook on an organization watches every
        extension in its namespace instead of one `id`, so a publish, yank,
        deprecation, rejection or owner change under `@{namespace}` reaches it
        without registering anything per extension. The per-extension webhooks
        at `/@{namespace}/{id}/webhooks` are a separate collection and are
        unaffected; the two do not see each other.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/WebhookInput' }
      responses:
        '201':
          description: Created, with the signing secret returned once
          content:
            application/json:
              schema: { $ref: '#/components/schemas/WebhookCreated' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '422': { $ref: '#/components/responses/UnprocessableEntity' }
  /orgs/{namespace}/webhooks/{webhookId}:
    parameters:
      - $ref: '#/components/parameters/NamespacePath'
      - name: webhookId
        in: path
        required: true
        schema: { type: integer }
    delete:
      summary: Delete an organization webhook
      operationId: deleteOrganizationWebhook
      tags: [Userspace]
      description: >
        Any owner, or an admin. An id belonging to a per-extension webhook is a
        `404` here, since the two collections do not overlap.
      responses:
        '204': { description: Deleted }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
...
  /extensions:
    get:
      summary: List published extensions
      operationId: listExtensions
      tags: [Open]
      security: []
      description: >
        Latest published version per `(namespace, id)`. Defaults to
        newest-first; `sort` changes the ranking. `namespace` narrows the
        listing to one account or organization, which is also what
        `GET /orgs/{namespace}/extensions` does with a membership check: an
        owner of the organization sees its private extensions, and everyone
        else sees only the public ones.
      parameters:
        - $ref: '#/components/parameters/Sort'
        - $ref: '#/components/parameters/LicenseFilter'
        - $ref: '#/components/parameters/NamespaceFilter'
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Dir'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Paginated extensions
          content:
            application/json:
              schema:
                type: object
                required: [data, _links]
                properties:
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/ExtensionSummary' }
                  _links: { $ref: '#/components/schemas/PageLinks' }
        '400': { $ref: '#/components/responses/BadRequest' }
...
components:
  parameters:
...
    NamespacePath:
      name: namespace
      in: path
      required: true
      schema: { $ref: '#/components/schemas/Namespace' }
...
  schemas:
    Namespace:
      type: string
      pattern: '^[a-z0-9](?:[a-z0-9-]{0,38}[a-z0-9])?$'
...
    Organization:
      type: object
      required: [namespace, displayName, bio, createdAt, _links]
      properties:
        namespace: { $ref: '#/components/schemas/Namespace' }
        displayName: { type: string }
        bio: { type: string }
        website: { type: ['string', 'null'], format: uri }
        github: { type: ['string', 'null'] }
        avatarUrl:
          type: ['string', 'null']
          format: uri
          description: >
            The external reference, or this organization's own
            `/orgs/{namespace}/avatar` path when the image was uploaded rather
            than referenced. `null` when there is neither, in which case the
            identicon is what `GET /orgs/{namespace}/avatar` serves.
        bannerUrl:
          type: ['string', 'null']
          format: uri
          description: >
            The external reference, or this organization's own
            `/orgs/{namespace}/banner` path after an upload.
        createdAt: { type: string, format: date-time, readOnly: true }
        _links:
          type: object
          required: [self, extensions, owners, avatar, banner]
          properties:
            self: { type: string }
            extensions: { type: string }
            owners: { type: string }
            avatar: { type: string }
            banner: { type: string }
      description: >
        An organization: a pseudo-account that owns a namespace, a profile and
        an extension collection, with no password, no sessions and no role. Its
        owners are accounts, and every owner has the same rights over it. A
        `kind` would be constant here, so the field is only on the account view,
        `GET /users/{namespace}`.
...
  responses:
...
    Unauthorized:
      description: Missing or invalid bearer token
      content: { application/problem+json: { schema: { $ref: '#/components/schemas/Problem' } } }
    Forbidden:
      description: >
        Authenticated, but not allowed to perform this action — e.g. not the
        resource owner or an admin, a namespace mismatch on publish, the
        current Terms of Service haven't been accepted yet
        (see `termsAcceptedVersion` on `PATCH /users/{namespace}`), or a
        credential missing the scope the route needs. The `detail` names the
        scope. An automation token on `/admin/*` or a version review also needs
        the `admin` scope on top of the account being an admin.
      content: { application/problem+json: { schema: { $ref: '#/components/schemas/Problem' } } }
...
    Conflict:
      description: Resource already exists
      content: { application/problem+json: { schema: { $ref: '#/components/schemas/Problem' } } }
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer